fix(error): consolidate duplicate captureError implementations (#52)
Remove captureError from lib/sentry.ts and consolidate all error
capturing in lib/error.ts. The lib/error.ts implementation now checks
isSentryEnabled before calling Sentry, providing consistent behavior.

- Remove captureError export from lib/sentry.ts
- Add isSentryEnabled guard to lib/error.ts
- Update useSharePoem hook to import from @/lib/error
- Update tests to reflect new module structure

closes #52
